+++ app/src/main/java/de/astronarren/allsky/data/location/LocationManager.kt
@@ -1,0 +1,1 @@
+// Move from utils to data/location package